Victor

+++ backend/src/main/java/br/com/todolistapp/todolist/model/Tarefa.java
@@ -8,20 +8,19 @@ public class Tarefa {
     @Id
     @Column
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     @Column
     private String descricao;
 
-
     @Column
     private boolean statusTarefa;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -40,6 +39,5 @@ public class Tarefa {
     public void setStatusTarefa(boolean statusTarefa) {
         this.statusTarefa = statusTarefa;
     }
-
-
+    
 }
